feat: convert history sync to durable workflows
